[#32] 모듈 분리
