--- feature/playlist/data/src/main/java/com/mashup/kkyuni/playlist/data/di/PlayListDataModule.kt
@@ -1,34 +1,0 @@
-package com.mashup.kkyuni.playlist.data.di
-
-import com.mashup.kkyuni.playlist.data.repository.PlayListRepository
-import com.mashup.kkyuni.playlist.data.repository.impl.PlayListRepositoryImpl
-import com.mashup.kkyuni.playlist.data.service.PlayListService
-import dagger.Binds
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ViewModelComponent
-import dagger.hilt.android.scopes.ViewModelScoped
-import dagger.hilt.components.SingletonComponent
-import retrofit2.Retrofit
-import javax.inject.Named
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-class PlayListDataModule {
-    @Provides
-    @Singleton
-    fun provideVideoService(@Named("kkyuni_api") retrofit: Retrofit): PlayListService =
-        retrofit.create(PlayListService::class.java)
-
-    @Module
-    @InstallIn(ViewModelComponent::class)
-    interface RepositoryModule {
-        @Binds
-        @ViewModelScoped
-        fun bindPlayListRepository(
-            repositoryImpl: PlayListRepositoryImpl
-        ): PlayListRepository
-    }
-}
